deps: upgrade storybook to v6.1
