--- .storybook/addons.js
@@ -1,4 +1,0 @@
-import "@storybook/addon-actions/register";
-import "@storybook/addon-links/register";
-import "@storybook/addon-knobs/register";
-import "storybook-addon-styled-component-theme/dist/register";
