+++ package.json
@@ -25,14 +25,15 @@
     "@babel/preset-flow": "7.0.0",
     "@babel/preset-react": "7.0.0",
     "@fortawesome/fontawesome-free": "^5.10.2",
-    "@storybook/addon-actions": "^5.2.8",
-    "@storybook/addon-info": "^5.2.8",
-    "@storybook/addon-knobs": "^5.2.8",
-    "@storybook/addon-links": "^5.2.8",
-    "@storybook/addon-storyshots": "^5.2.8",
-    "@storybook/react": "^5.2.8",
-    "@storybook/storybook-deployer": "^2.8.1",
-    "@storybook/theming": "^5.2.8",
+    "@storybook/addon-actions": "^6.1.18",
+    "@storybook/addon-essentials": "^6.1.18",
+    "@storybook/addon-info": "^5.3.21",
+    "@storybook/addon-knobs": "^6.1.18",
+    "@storybook/addon-links": "^6.1.18",
+    "@storybook/addon-storyshots": "^6.1.18",
+    "@storybook/react": "^6.1.18",
+    "@storybook/storybook-deployer": "^2.8.7",
+    "@storybook/theming": "^6.1.18",
     "@typescript-eslint/eslint-plugin": "^2.3.0",
     "@typescript-eslint/parser": "^2.3.0",
     "babel-eslint": "^10.0.3",
@@ -58,7 +59,7 @@
     "react-virtualized": "^9.21.1",
     "rimraf": "^3.0.0",
     "source-map-explorer": "^2.0.1",
-    "storybook-addon-styled-component-theme": "^1.2.4",
+    "storybook-addon-styled-component-theme": "^1.3.0",
     "styled-components": "^4.3.2",
     "vega": "^5.17.3",
     "vega-embed": "^6.0.0",
